requirements.txt
json5==0.9.14
paho_mqtt==1.6.1
pyserial==3.5
